Distribution, mean, stdDev added

+++ SupplyChain/dserver.js
@@ -38,9 +38,17 @@ app.listen(process.env.PORT || 3000, function(){
     console.log('Your node js server is running');
 });
 
-app.post('/deployGameWithFile', upload.single('uploadCsv'), deployGameWithFile);
+app.get('/gameInfo', function(req, res) {
+	fs.readFile('gameInfo.json', (err, data)=> {
+		if (err) throw err;
+    var gameInfo = JSON.parse(data);
+    console.log(gameInfo);
+    res.status(200).send(gameInfo);
+	});
+});
 
-app.get('/getPlayers', getPlayers);
+
+app.post('/deployGameWithFile', upload.single('uploadCsv'), deployGameWithFile);
 
 
 var players = {retailer: {}, wholesaler: {}, distributer: {}, factory: {}};
@@ -50,15 +58,22 @@ function deployGameWithFile(req, res) {
 
 console.log("\n\n\n");
   const fileRows = [];
-  csv.parseFile(req.file.path)
-    .on("data", function (data) {
-      fileRows.push(data); // push each row
-    })
-    .on("end", function () {
-      console.log(fileRows)
-      fs.unlinkSync(req.file.path);   // remove temp file
-      //process "fileRows" and respond
-    });
+  if(req.file){
+  	//file is uploaded
+  	csv.parseFile(req.file.path)
+    	.on("data", function (data) {
+      	fileRows.push(data); // push each row
+    	})
+    	.on("end", function () {
+      	console.log(fileRows)
+      	fs.unlinkSync(req.file.path);   // remove temp file
+     	 //process "fileRows" and respond
+    	});
+
+	}
+	else {
+		//file not found
+	}
     console.log("\n");
     console.log(req.body);
     console.log("\n");
@@ -83,7 +98,18 @@ console.log("\n\n\n");
 	console.log("Number of weeks: " + totalWeeks);
 	console.log(hCost)
 	
+	var distribution = req.body.distribution;
+	var stdDev = Number(req.body.stdDev);
+	var mean = Number(req.body.mean);
 
+	var gameInfo = {};
+	gameInfo.distribution = distribution;
+	gameInfo.stdDev = stdDev;
+	gameInfo.mean = mean;
+	console.log(gameInfo);
+
+	var data = JSON.stringify(gameInfo);
+	fs.writeFileSync('gameInfo.json', data);
 //retailer account
 web3.eth.personal.newAccount("pass1").then(function(newAcc) {
 	players.retailer["address"] = newAcc;
@@ -145,15 +171,11 @@ myContract.deploy({
 	web3.eth.personal.lockAccount(accounts[len-4]);
 
 	console.log("Success");
-
+	console.log(players);
+	res.status(200).send(players);
 });
 });
 });});});});
 
-	res.status(200).send(players);
-}
-
-function getPlayers(req, res){
-	console.log(players);
-	res.status(200).send(players);
+	
 }
